fix all the build warnings

+++ src/allocator/fy-allocator-dedup.c
@@ -142,10 +142,10 @@ static int fy_dedup_tag_data_setup(struct fy_dedup_tag_data *dtd,
 			dtd->chain_length_grow_trigger++;
 	}
 
-	dtd->bloom_filter_mask = (1U << dtd->bloom_filter_bits) - 1;
-	dtd->bucket_count_mask = (1U << dtd->bucket_count_bits) - 1;
+	dtd->bloom_filter_mask = ((unsigned int)1 << dtd->bloom_filter_bits) - 1;
+	dtd->bucket_count_mask = ((unsigned int)1 << dtd->bucket_count_bits) - 1;
 
-	dtd->bloom_id_count = ((1U << dtd->bloom_filter_bits) + FY_ID_BITS_BITS - 1) / FY_ID_BITS_BITS;
+	dtd->bloom_id_count = (((size_t)1 << dtd->bloom_filter_bits) + FY_ID_BITS_BITS - 1) / FY_ID_BITS_BITS;
 	assert(dtd->bloom_id_count > 0);
 	tmpsz = dtd->bloom_id_count * sizeof(*dtd->bloom_id);
 	dtd->bloom_id = fy_parent_allocator_alloc(&da->a, tmpsz, _Alignof(fy_id_bits));
@@ -153,7 +153,7 @@ static int fy_dedup_tag_data_setup(struct fy_dedup_tag_data *dtd,
 		goto err_out;
 	fy_id_reset(dtd->bloom_id, dtd->bloom_id_count);
 
-	dtd->bucket_count = 1U << dtd->bucket_count_bits;
+	dtd->bucket_count = (size_t)1 << dtd->bucket_count_bits;
 	assert(dtd->bucket_count);
 
 	tmpsz = sizeof(*dtd->buckets) * dtd->bucket_count;
@@ -163,7 +163,7 @@ static int fy_dedup_tag_data_setup(struct fy_dedup_tag_data *dtd,
 	for (i = 0; i < dtd->bucket_count; i++)
 		atomic_store(&dtd->buckets[i], NULL);
 
-	dtd->bucket_id_count = ((1U << dtd->bucket_count_bits) + FY_ID_BITS_BITS - 1) / FY_ID_BITS_BITS;
+	dtd->bucket_id_count = (((size_t)1 << dtd->bucket_count_bits) + FY_ID_BITS_BITS - 1) / FY_ID_BITS_BITS;
 	assert(dtd->bucket_id_count > 0);
 	tmpsz = dtd->bucket_id_count * sizeof(*dtd->buckets_in_use);
 	dtd->buckets_in_use = fy_parent_allocator_alloc(&da->a, tmpsz, _Alignof(fy_id_bits));
@@ -515,7 +515,7 @@ static int fy_dedup_setup(struct fy_allocator *a, struct fy_allocator *parent, i
 	bucket_count_bits = cfg->bucket_count_bits;
 	if (!bucket_count_bits && has_estimate) {
 		bucket_count_bits = 1;
-		while ((1LU << bucket_count_bits) < (cfg->estimated_content_size / BUCKET_ESTIMATE_DIV))
+		while (((size_t)1 << bucket_count_bits) < (cfg->estimated_content_size / BUCKET_ESTIMATE_DIV))
 			bucket_count_bits++;
 #ifdef DEBUG_GROWS
 		fprintf(stderr, "bucket_count_bits %u\n", bucket_count_bits);
@@ -531,7 +531,7 @@ static int fy_dedup_setup(struct fy_allocator *a, struct fy_allocator *parent, i
 	bloom_filter_bits = cfg->bloom_filter_bits;
 	if (!bloom_filter_bits && has_estimate) {
 		bloom_filter_bits = 1;
-		while ((1LU << bloom_filter_bits) < (cfg->estimated_content_size / BLOOM_ESTIMATE_DIV))
+		while (((size_t)1 << bloom_filter_bits) < (cfg->estimated_content_size / BLOOM_ESTIMATE_DIV))
 			bloom_filter_bits++;
 #ifdef DEBUG_GROWS
 		fprintf(stderr, "bloom_filter_bits %u\n", bloom_filter_bits);
